Port more tests
